+++ src/exception/ManagerLoadException.java
@@ -1,4 +1,4 @@
-package manager;
+package exception;
 
 public class ManagerLoadException extends RuntimeException {
