+++ src/exception/ManagerSaveException.java
@@ -1,4 +1,4 @@
-package manager;
+package exception;
 
 public class ManagerSaveException extends RuntimeException {
