+++ src/exception/TaskIntersectionException.java
@@ -1,14 +1,15 @@
-package task;
+package exception;
 
-public class WrongCSVLineException extends Exception {
+public class TaskIntersectionException extends Exception {
 
     /**
      * Constructs a new exception with {@code null} as its detail message.
      * The cause is not initialized, and may subsequently be initialized by a
      * call to {@link #initCause}.
      */
-    public WrongCSVLineException() {
+    public TaskIntersectionException() {
     }
+
 
     /**
      * Constructs a new exception with the specified detail message.  The
@@ -18,7 +19,8 @@ public class WrongCSVLineException extends Exception {
      * @param message the detail message. The detail message is saved for
      *                later retrieval by the {@link #getMessage()} method.
      */
-    public WrongCSVLineException(String message) {
+    public TaskIntersectionException(String message) {
         super(message);
     }
+
 }
