+++ src/exception/TaskNotFoundException.java
@@ -1,4 +1,4 @@
-package manager;
+package exception;
 
 public class TaskNotFoundException extends Exception {
   /**
