+++ src/exception/WrongCSVLineException.java
@@ -1,12 +1,13 @@
-package manager;
+package exception;
 
-public class WrongTaskArgumentException extends Exception {
+public class WrongCSVLineException extends Exception {
+
     /**
      * Constructs a new exception with {@code null} as its detail message.
      * The cause is not initialized, and may subsequently be initialized by a
      * call to {@link #initCause}.
      */
-    public WrongTaskArgumentException() {
+    public WrongCSVLineException() {
     }
 
     /**
@@ -17,7 +18,7 @@ public class WrongTaskArgumentException extends Exception {
      * @param message the detail message. The detail message is saved for
      *                later retrieval by the {@link #getMessage()} method.
      */
-    public WrongTaskArgumentException(String message) {
+    public WrongCSVLineException(String message) {
         super(message);
     }
 }
