+++ src/exception/WrongTaskArgumentException.java
@@ -1,15 +1,13 @@
-package manager;
+package exception;
 
-public class TaskIntersectionException extends Exception {
-
+public class WrongTaskArgumentException extends Exception {
     /**
      * Constructs a new exception with {@code null} as its detail message.
      * The cause is not initialized, and may subsequently be initialized by a
      * call to {@link #initCause}.
      */
-    public TaskIntersectionException() {
+    public WrongTaskArgumentException() {
     }
-
 
     /**
      * Constructs a new exception with the specified detail message.  The
@@ -19,8 +17,7 @@ public class TaskIntersectionException extends Exception {
      * @param message the detail message. The detail message is saved for
      *                later retrieval by the {@link #getMessage()} method.
      */
-    public TaskIntersectionException(String message) {
+    public WrongTaskArgumentException(String message) {
         super(message);
     }
-
 }
