+++ src/http/deserialiazer/DurationTypeAdapter.java
@@ -1,4 +1,4 @@
-package http;
+package http.deserialiazer;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
