+++ src/http/deserialiazer/TaskListTypeToken.java
@@ -1,4 +1,4 @@
-package http;
+package http.deserialiazer;
 
 import com.google.gson.reflect.TypeToken;
 import task.Task;
