+++ src/manager/InMemoryTaskManager.java
@@ -1,5 +1,12 @@
 package manager;
-import task.*;
+
+import exception.TaskIntersectionException;
+import exception.TaskNotFoundException;
+import exception.WrongTaskArgumentException;
+import task.Epic;
+import task.Subtask;
+import task.Task;
+import task.TaskFactory;
 
 import java.util.*;
 import java.util.stream.Collectors;
