+++ test/manager/TaskManagerTest.java
@@ -1,5 +1,8 @@
 package manager;
 
+import exception.TaskIntersectionException;
+import exception.TaskNotFoundException;
+import exception.WrongTaskArgumentException;
 import org.junit.jupiter.api.Test;
 import task.Epic;
 import task.Subtask;
