+++ test/task/TaskTest.java
@@ -1,8 +1,10 @@
 package task;
 
 import org.junit.jupiter.api.Test;
+
 import java.time.Duration;
 import java.time.LocalDateTime;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class TaskTest {
